Get the ONLY_LIGHT_CONE to work

+++ node/application/wsnLogic/WsnLogicPacket_m.h
@@ -28,7 +28,7 @@ struct WsnLogicData
     unsigned short type;
     unsigned short originNodeID;
     unsigned short senderNodeID;
-    unsigned short destinationNodesID;
+    int destinationNodesID;
     int hop;
 };
